+++ pages/index.js
@@ -2,8 +2,16 @@ import Head from 'next/head'
 import { BiUserPlus } from "react-icons/bi";
 import Table from '../components/table';
 import Form from '../components/form';
+import { useState } from 'react';
 
 export default function Home() {
+
+  const [visible, setVisible] = useState(false)
+
+  const handler = () => {
+    setVisible(!visible)
+  }
+
   return (
     <section>
       <Head>
@@ -17,16 +25,14 @@ export default function Home() {
 
         <div className="container mx-auto flex justify-between py-5 border-b">
             <div className="left flex gap-3">
-                <button className='flex bg-indigo-500 text-white px-4 py-2 border rounded-md hover:bg-grary-50 hover:border-indigo-500 hover:text-gray-800'>
+                <button onClick={handler} className='flex bg-indigo-500 text-white px-4 py-2 border rounded-md hover:bg-grary-50 hover:border-indigo-500 hover:text-gray-800'>
                   Add Employee <span className='px-1'><BiUserPlus size={23}></BiUserPlus></span>
                 </button>
             </div>
         </div>
 
         {/* collapsable form */}
-        <div className="container mx-auto py-5">
-          <Form></Form>
-        </div>
+        { visible ? <Form></Form> : <></>}
 
         {/* table */}
         <div className="container mx-auto">
